created smart enemy.

+++ src/com/tutorial/main/FastEnemy.java
@@ -7,7 +7,7 @@ import java.awt.Rectangle;
 public class FastEnemy extends GameObject {
 	private Handler handler;
 	
-	public FastEnemy(int x, int y, ID id, Handler handler) {
+	public FastEnemy(float x, float y, ID id, Handler handler) {
 		super(x, y, id);
 		
 		this.handler = handler;
@@ -17,7 +17,7 @@ public class FastEnemy extends GameObject {
 	}
 	
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	@Override
@@ -39,6 +39,6 @@ public class FastEnemy extends GameObject {
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.CYAN);		
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 }
